+++ SemanticUtil.h
@@ -10,10 +10,10 @@ void init_random();
 int random_0_to_n(int n);
 
 // Method used to calculate the factorial of some given number.
-int factorial(int n);
+unsigned long factorial(unsigned long n);
 
 // Method used to calculate the entropy of M equiprobable choices.
-double entropyEquiprobable(int m);
+double entropyEquiprobable(long m);
 
 // Method used to write JSON to a give file.
 void writeJSONToFile(std::string outputPath, int version, std::string fileName, Json::Value output);
